@@ -2,19 +2,22 @@ apply plugin: 'com.android.application'
 
 apply plugin: 'kotlin-android'
 
-apply plugin: 'kotlin-android-extensions'
-
 apply plugin: 'kotlin-kapt'
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 33
+
+    buildFeatures {
+        viewBinding true
+    }
+
     defaultConfig {
         applicationId "com.ocram.qichwadic"
         minSdkVersion 19
-        targetSdkVersion 29
-        versionCode 21
-        versionName "3.2.0"
+        targetSdkVersion 33
+        versionCode 22
+        versionName "3.3.0"
+        multiDexEnabled true
         testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         javaCompileOptions {
             annotationProcessorOptions {
@@ -44,21 +47,24 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
 
     kotlinOptions {
-        jvmTarget = JavaVersion.VERSION_1_8
+        jvmTarget = JavaVersion.VERSION_17
     }
 
     sourceSets {
         androidTest.assets.srcDirs += files("$projectDir/schemas".toString())
     }
-
     packagingOptions {
-        exclude 'META-INF/atomicfu.kotlin_module'
+        resources {
+            excludes += ['META-INF/atomicfu.kotlin_module']
+        }
     }
+    namespace 'com.ocram.qichwadic'
+
 }
 
 dependencies {
@@ -74,13 +80,11 @@ dependencies {
     implementation "androidx.constraintlayout:constraintlayout:$constraint_layout"
     implementation "androidx.cardview:cardview:$cardview"
     implementation "com.google.android.material:material:$material"
-//    implementation "androidx.legacy:legacy-preference-v14:$preferencev14"
 
     // Architecture components
     // room
     implementation "androidx.room:room-runtime:$roomVersion"
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-//    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation "androidx.appcompat:appcompat:$appcompat"
     kapt "androidx.room:room-compiler:$roomVersion"
     implementation "androidx.room:room-ktx:$roomVersion"
 
@@ -88,18 +92,16 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-extensions:$archLifecycleExtensionsVersion"
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$archLifecycleVersion"
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:$archLifecycleVersion"
-    implementation "androidx.lifecycle:lifecycle-reactivestreams:$archLifecycleVersion"
+    implementation "androidx.lifecycle:lifecycle-reactivestreams-ktx:$archLifecycleVersion"
 
     // Navigation
     implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
     implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
 
-
     // Retrofit / LeakCanary
     implementation "com.squareup.retrofit2:retrofit:$retrofit2"
     implementation "com.squareup.retrofit2:converter-gson:$retrofit2gson"
     debugImplementation "com.squareup.leakcanary:leakcanary-android:$leakCanary"
-
 
     // Testing
     testImplementation "junit:junit:$junit"
@@ -111,13 +113,10 @@ dependencies {
     implementation "com.github.AppIntro:AppIntro:$appIntro"
 
     // Resolve conflicts between main and test APK:
-    androidTestImplementation "androidx.annotation:annotation:$annotation"
+    implementation "androidx.annotation:annotation:$annotation"
     androidTestImplementation "androidx.appcompat:appcompat:$appcompat"
     androidTestImplementation "com.google.android.material:material:$material"
 
-    // Koin AndroidX Scope feature
-    implementation "org.koin:koin-androidx-scope:$koin_version"
-    // Koin AndroidX ViewModel feature
-    implementation "org.koin:koin-androidx-viewmodel:$koin_version"
-
+    // koin
+    implementation "io.insert-koin:koin-android:$koin_version"
 }
