@@ -1,12 +1,12 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
-    ext.kotlin_version = '1.3.72'
+    ext.kotlin_version = '1.8.20'
     repositories {
-        jcenter()
+        mavenCentral()
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:7.0.2'
+        classpath 'com.android.tools.build:gradle:8.0.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -15,7 +15,7 @@ buildscript {
 
 allprojects {
     repositories {
-        jcenter()
+        mavenCentral()
         // required to find the project's artifacts
         maven { url 'https://maven.google.com' }
         maven { url 'https://jitpack.io' }
@@ -28,25 +28,25 @@ task clean(type: Delete) {
 }
 
 ext {
-    coroutines = '1.3.4'
-    corektx = '1.3.2'
-    koin_version= '2.1.5'
-    appcompat = '1.1.0'
-    constraint_layout = '1.1.3'
+    coroutines = '1.6.4'
+    corektx = '1.10.0'
+    koin_version= '3.4.0'
+    appcompat = '1.6.1'
+    constraint_layout = '2.1.4'
     cardview = '1.0.0'
-    material = '1.3.0'
+    material = '1.8.0'
     preferencev14 = '1.0.0'
     appIntro = '6.1.0'
-    roomVersion = '2.2.6'
+    roomVersion = '2.5.1'
     archLifecycleExtensionsVersion = '2.2.0'
-    archLifecycleVersion = '2.3.1'
-    nav_version = "2.3.4"
+    archLifecycleVersion = '2.6.1'
+    nav_version = "2.5.3"
     retrofit2 = '2.9.0'
     retrofit2gson = '2.9.0'
-    leakCanary = '2.4'
-    junit = '4.13'
-    extjunit = '1.1.1'
-    testrunner = '1.2.0'
-    annotation = '1.1.0'
+    leakCanary = '2.10'
+    junit = '4.13.2'
+    extjunit = '1.1.5'
+    testrunner = '1.5.2'
+    annotation = '1.6.0'
     debugdb = '1.0.6'
 }
