@@ -21,10 +21,13 @@ import com.ocram.qichwadic.BuildConfig
 
 import com.ocram.qichwadic.R
 import com.ocram.qichwadic.core.ui.activity.MainActivity
-import kotlinx.android.synthetic.main.fragment_about.*
+import com.ocram.qichwadic.databinding.FragmentAboutBinding
 import java.util.*
 
 class AboutFragment : Fragment() {
+
+    private var _binding: FragmentAboutBinding? = null;
+    private val binding get() = _binding!!
 
     private fun TextView.removeLinksUnderline() {
         val spannable = SpannableString(text)
@@ -40,8 +43,9 @@ class AboutFragment : Fragment() {
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_about, container, false)
+                              savedInstanceState: Bundle?): View {
+        _binding = FragmentAboutBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -54,23 +58,23 @@ class AboutFragment : Fragment() {
 
         toolbar.setupWithNavController(navController, appBarConfiguration)
 
-        tvProjectCollab.text = HtmlCompat.fromHtml(getString(R.string.about_project_collaboration), HtmlCompat.FROM_HTML_MODE_LEGACY)
-        tvProjectCollab.removeLinksUnderline()
-        tvProjectCollab.movementMethod = LinkMovementMethod.getInstance()
+        binding.tvProjectCollab.text = HtmlCompat.fromHtml(getString(R.string.about_project_collaboration), HtmlCompat.FROM_HTML_MODE_LEGACY)
+        binding.tvProjectCollab.removeLinksUnderline()
+        binding.tvProjectCollab.movementMethod = LinkMovementMethod.getInstance()
 
-        tvCopyright.text = getString(R.string.about_copyright, Calendar.getInstance().get(Calendar.YEAR))
+        binding.tvCopyright.text = getString(R.string.about_copyright, Calendar.getInstance().get(Calendar.YEAR))
         bindEvents()
     }
 
     private fun bindEvents() {
-        btnRateApp.setOnClickListener {
+        binding.btnRateApp.setOnClickListener {
             (requireActivity() as MainActivity).openMarketIntent()
         }
 
         val qichwaSiteUrl = "https://www.qichwa.net"
-        btnWebsite.setOnClickListener { (requireActivity() as MainActivity).openActionViewIntent(qichwaSiteUrl) }
+        binding.btnWebsite.setOnClickListener { (requireActivity() as MainActivity).openActionViewIntent(qichwaSiteUrl) }
 
-        btnContact.setOnClickListener {
+        binding.btnContact.setOnClickListener {
             (requireActivity() as MainActivity)
                     .openEmailIntent(
                             BuildConfig.DEV_EMAIL,
@@ -78,7 +82,7 @@ class AboutFragment : Fragment() {
                     )
         }
 
-        btnYoutube.setOnClickListener {
+        binding.btnYoutube.setOnClickListener {
             val qichwa20YtUrl = "https://www.youtube.com/channel/UCZ5kIwvo7DlN9qdrQrjUOkg"
             (requireActivity() as MainActivity)
                     .openActionViewIntent(qichwa20YtUrl)
