@@ -10,7 +10,6 @@ import android.widget.ArrayAdapter
 import androidx.appcompat.widget.Toolbar
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupWithNavController
@@ -20,21 +19,26 @@ import com.google.android.material.snackbar.Snackbar
 import com.ocram.qichwadic.R
 import com.ocram.qichwadic.core.domain.model.DictionaryModel
 import com.ocram.qichwadic.core.ui.DictLang
-import kotlinx.android.synthetic.main.fragment_dictionary.*
+import com.ocram.qichwadic.databinding.FragmentDictionaryBinding
 import org.koin.androidx.viewmodel.ext.android.viewModel
-import java.util.*
 
 class DictionaryFragment : Fragment(), AdapterView.OnItemSelectedListener, DictionaryAdapter.DefinitionDownloadListener {
 
     private val dictionaryViewModel: DictionaryViewModel by viewModel()
+    private var _binding: FragmentDictionaryBinding? = null
+    private val binding get() = _binding!!
+
     private lateinit var dictionaryAdapter: DictionaryAdapter
     private lateinit var dictLangMap: Map<String, List<DictionaryModel>>
     private lateinit var dictLangs: List<DictLang>
     private var currentLangCode: String? = null
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_dictionary, container, false)
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentDictionaryBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -47,10 +51,15 @@ class DictionaryFragment : Fragment(), AdapterView.OnItemSelectedListener, Dicti
         toolbar.setupWithNavController(navController, appBarConfiguration)
 
         initViews()
-        dictionaryViewModel.dictionariesByLang.observe(this, Observer { this.onDictionaryListChanged(it) })
-        dictionaryViewModel.localLoading.observe(this, Observer { this.onLocalDictionariesLoadingChanged(it) })
-        dictionaryViewModel.dictionaryActionStatus.observe(this, Observer { this.onDictionaryAction(it) })
-        dictionaryViewModel.cloudError.observe(this, Observer {  this.onCloudError(it) })
+        dictionaryViewModel.dictionariesByLang.observe(viewLifecycleOwner) { this.onDictionaryListChanged(it) }
+        dictionaryViewModel.localLoading.observe(viewLifecycleOwner) { this.onLocalDictionariesLoadingChanged(it) }
+        dictionaryViewModel.dictionaryActionStatus.observe(viewLifecycleOwner) { this.onDictionaryAction(it) }
+        dictionaryViewModel.cloudError.observe(viewLifecycleOwner) { this.onCloudError(it) }
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
     }
 
     private fun initViews() {
@@ -58,29 +67,31 @@ class DictionaryFragment : Fragment(), AdapterView.OnItemSelectedListener, Dicti
 
         setRecyclerView()
         setSpinnerAdapters()
-        spTargetLanguages.onItemSelectedListener = this
+
+        binding.spTargetLanguages.onItemSelectedListener = this
     }
 
     private fun setRecyclerView() {
-        dictionaryAdapter = DictionaryAdapter(ArrayList(), this)
+        dictionaryAdapter = DictionaryAdapter(mutableListOf(), this)
+
         val linearLayoutManager = LinearLayoutManager(requireContext(), RecyclerView.VERTICAL, false)
-        rvDictionaries.layoutManager = linearLayoutManager
-        rvDictionaries.isNestedScrollingEnabled = false
-        rvDictionaries.adapter = dictionaryAdapter
+
+        binding.rvDictionaries.layoutManager = linearLayoutManager
+        binding.rvDictionaries.isNestedScrollingEnabled = false
+        binding.rvDictionaries.adapter = dictionaryAdapter
     }
 
     private fun setSpinnerAdapters() {
         val langAdapter = ArrayAdapter(requireContext(), R.layout.item_spinner_dict_lang_white, dictLangs)
         langAdapter.setDropDownViewResource(R.layout.item_spinner_dict_lang_white)
-        spTargetLanguages.adapter = langAdapter
-        spTargetLanguages.setSelection(0)
+        binding.spTargetLanguages.adapter = langAdapter
+        binding.spTargetLanguages.setSelection(0)
     }
 
     override fun onItemSelected(p0: AdapterView<*>?, p1: View?, pos: Int, p3: Long) {
-        val dictLang = spTargetLanguages.getItemAtPosition(pos) as DictLang
+        val dictLang = binding.spTargetLanguages.getItemAtPosition(pos) as DictLang
         this.currentLangCode = dictLang.code
-        dictionaryAdapter.dictionaries = dictLangMap[dictLang.code]
-        dictionaryAdapter.notifyDataSetChanged()
+        dictionaryAdapter.updateDictionaries(dictLangMap[dictLang.code] ?: emptyList())
     }
 
     override fun onNothingSelected(p0: AdapterView<*>?) {}
@@ -96,39 +107,34 @@ class DictionaryFragment : Fragment(), AdapterView.OnItemSelectedListener, Dicti
     private fun onDictionaryListChanged(dictionariesMap: Map<String, List<DictionaryModel>>) {
         this.dictLangMap = dictionariesMap
         if (dictionariesMap.isNotEmpty()) {
-            showDictionaries(dictionariesMap[this.currentLangCode])
+            showDictionaries(dictionariesMap[this.currentLangCode] ?: emptyList())
         }
     }
 
-    private fun showDictionaries(dictionaries: List<DictionaryModel>?) {
-        dictionaryAdapter.dictionaries = dictionaries
-        dictionaryAdapter.notifyDataSetChanged()
-        spTargetLanguages.visibility = View.VISIBLE
-        rvDictionaries.visibility = View.VISIBLE
+    private fun showDictionaries(dictionaries: List<DictionaryModel>) {
+        dictionaryAdapter.updateDictionaries(dictionaries)
+        binding.spTargetLanguages.visibility = View.VISIBLE
+        binding.rvDictionaries.visibility = View.VISIBLE
     }
 
     private fun onLocalDictionariesLoadingChanged(isLoading: Boolean?) {
         if (isLoading != null) {
-            pbDictionariesLoading.visibility = if (isLoading) View.VISIBLE else View.GONE
+            binding.pbDictionariesLoading.visibility = if (isLoading) View.VISIBLE else View.GONE
         }
     }
 
     private fun onDictionaryAction(dictionaryActionState: DictionaryActionState) {
 
-        val messageTemplate: Int?
-
-        if(dictionaryActionState.error) {
-            messageTemplate =
-                    if(dictionaryActionState.dictionary.existsInLocal)
-                        R.string.dictionary_delete_error
-                    else
-                        R.string.dictionary_save_error
+        val messageTemplate: Int = if(dictionaryActionState.error) {
+            if(dictionaryActionState.dictionary.existsInLocal)
+                R.string.dictionary_delete_error
+            else
+                R.string.dictionary_save_error
         } else {
-            messageTemplate =
-                    if(dictionaryActionState.dictionary.existsInLocal)
-                        R.string.dictionary_save_success
-                    else
-                        R.string.dictionary_delete_success
+            if(dictionaryActionState.dictionary.existsInLocal)
+                R.string.dictionary_save_success
+            else
+                R.string.dictionary_delete_success
         }
         showMessage(getString(messageTemplate, dictionaryActionState.dictionary.name))
     }
@@ -141,7 +147,7 @@ class DictionaryFragment : Fragment(), AdapterView.OnItemSelectedListener, Dicti
 
     private fun showMessage(message: String?) {
         if (!message.isNullOrEmpty()) {
-            Snackbar.make(clDictionaries, message, Snackbar.LENGTH_SHORT).show()
+            Snackbar.make(binding.clDictionaries, message, Snackbar.LENGTH_SHORT).show()
         }
     }
 }
