@@ -1,15 +1,14 @@
 package com.ocram.qichwadic.features.search.ui
 
-import android.annotation.SuppressLint
 import android.content.Context
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ArrayAdapter
 import android.widget.TextView
-
 import com.ocram.qichwadic.R
 import com.ocram.qichwadic.core.ui.DictLang
+import java.util.*
 
 class SearchDictLangAdapter(context: Context) : ArrayAdapter<DictLang>(context, R.layout.item_spinner_dict_lang) {
 
@@ -17,7 +16,7 @@ class SearchDictLangAdapter(context: Context) : ArrayAdapter<DictLang>(context, 
 
     init {
         val codeTextLangs = context.resources.getStringArray(R.array.dictLangs)
-        this.dictLangs = codeTextLangs.map { DictLang(it) }.filter { !it.isQuechua() }.map { it }
+        this.dictLangs = codeTextLangs.map { DictLang(it) }.filter { !it.isQuechua() }
     }
 
     override fun getCount(): Int {
@@ -34,11 +33,10 @@ class SearchDictLangAdapter(context: Context) : ArrayAdapter<DictLang>(context, 
         return -1
     }
 
-    override fun getItem(position: Int): DictLang? {
+    override fun getItem(position: Int): DictLang {
         return dictLangs[position]
     }
 
-    @SuppressLint("DefaultLocale")
     private fun getCustomView(position: Int, parent: ViewGroup, dropdown: Boolean): View {
         val inflater = LayoutInflater.from(parent.context)
 
@@ -47,10 +45,8 @@ class SearchDictLangAdapter(context: Context) : ArrayAdapter<DictLang>(context, 
 
         val dictLang = getItem(position)
         val tvName = layout.findViewById<TextView>(R.id.tvName)
-        if (dictLang != null) {
-            val displayText = if (dropdown) dictLang.name else dictLang.code.toUpperCase()
-            tvName.text = displayText
-        }
+        val displayText = if (dropdown) dictLang.name else dictLang.code.uppercase(Locale.ROOT)
+        tvName.text = displayText
         return layout
     }
 
