@@ -7,7 +7,7 @@ abstract class CloudDataStore {
     fun <T> processResponse(response: Response<T>): ApiResponse<out T> {
         if(response.isSuccessful) {
             response.body()?.let {
-                return ApiResponse.Success(response.body()!!)
+                return ApiResponse.Success(it)
             }
         }
         return ApiResponse.Error(response.code())
